gray2binary() function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array, if the bit depth is less than 8 then we read and write the “colorTable”. Then we compare each pixel value with the threshold value which is 128, if the pixel value is greater than the threshold value, we modify it to store 255 else 0, then using “fwrite ()” function we write back the data into the output file.