color2gray () function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then we operate the loop to read the image data and the pixels are then modified using the conversion formula and then using the put function we write back the image data into the output file.