negative() function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then we use “malloc()” function to allocate the memory to store the image data then the pixels are accessed with two for loop and they are modified by subtracting it from 255 and stored in an other array , which then write back to the output file.