crop function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the size of the image the we assign the new height and width to the image header and the new location from where the printing of data will be start if the bit depth is less than 8 then we read and write the “colorTable”. then we simply operate the loop upto the new height and width value and print the image data in the output file. The process is similar for the RGB type of image.