flip() function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the size of the image then the statement is asked to flip horizontally or vertically to the user where the switch() function takes execution to the right case, let’s say it takes to the case 1 which is to flip horizontally as the execution proceeds the pixel of the input image is accessed with the loop and it then modified where the 1st row stores the last row data and the 2nd row of second last row data like this so on, and the image got flip then we write back it into the output file. Similarly, is with vertical flip but here in place of row, column as exchanged.