meangray () function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the size of the image if the bit depth is less than 8 then we read and write the “colorTable”. Then using the loop, we access the pixel where every pixel value is added to a new variable. Then after the loop ends the variable is divided with the size of the image to get the meangray value. Which further gets stored in the output text file.