morph() function
In this function first the input & two output image file is opened and checking is performed weather they are opened successfully or not. Then the image header of both the input image is accessed with the help of loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is declared that to which memory is statically allocated with size equal to the size of the image if the bit depth is less than 8 then we read and write the “colorTable”. Then using the loop, we read both image data and then the pixel of both the images are combined in a particular proportion value which is entered by the user and it get stored in the structure image variable which then is used to write back into the output file. The process is similar for the RGB type of image.