getpixel () function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the size of the image if the bit depth is less than 8 then we read the “colorTable”. Then we read the image data and then print the pixel data of the particular location according to the coordinates entered by the user into the output text file. The process is similar for the RGB type of image.