translate() function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the size of the image if the bit depth is less than 8 then we read and write the “colorTable”.Then we read the image data and using the loop we access the pixel data and modify it by adding the translate factor “t” and assign it to the coordinates of the variable of the structure image , then we simply print the structure image variable data back to the output file. The process is similar for the RGB type of the image.