rotate() function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. We then convert the degree value into the float type that will be use in the formula which will rotate the image. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the size of the image if the bit depth is less than 8 then we read and write the “colorTable”. Then using the loop, we access the input image pixel location and assign the value to the new coordinates using the formula which will rotate the image about the center and if the value of the coordinates exceeds the height and width of the image, we assign them 0 value. Then we store the new coordinate value in an array and writ back it into the output file. The process is similar for the RGB type of image.