Bright() function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the size of the image if the bit depth is less than 8 then we read and write the “colorTable”. then using the loop we access the pixel and modify it by adding some extra value i.e. BRIGHTNESS whose value assign by the use , then we check if the pixel value is greater than the  255  or not , if it is then we assign 255 to it , after this we write back the image data into the output file. The process is similar for the RGB type of image.