blur() function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the size of the image if the bit depth is less than 8 then we read and write the “colorTable”. Then we declare a float type 3X3 matrix that basically stores the blurring factor after it we use loop to access the pixel values and modify it by multiplying with the matrix and storing the value in another variable , that further stored in another array with the size equal to the image size and then we write back the array in the output file. The process is similar for the RGB type of image.