Sepiafilter () function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the image size then using the loop we modify the pixel which is constructed with Red ,Green ,Blue  color using the formula . then we write back the image data into the output file.