smooth() function
In this function first the input & output image file is opened and checking is performed weather they are opened successfully or not. Then the image header is accessed with the help of one for loop and it is stored inside an array ,then we write the header back to the output file then image height, width and bitDepth is extracted from input image Header then a variable is statically allocated with size equal to the size of the image if the bit depth is less than 8 then we read and write the “colorTable”. Then we read the image content, then we operate the loop upto the size of the image and replaces every pixel value by the mean of the neighboring pixel. Then we write back it into the output file. The process is similar for the RGB type of the image.